introduce the proto-MIR (PMIR)
It only covers simple expressions (no calls, constructions, and complex
expressions), and acts as an intermediary between the AST and the MIR,
moving most of the expression-related judgement and decision making
away from `mirgen`.

During production of the proto-MIR, there's also some lowering applied,
which is necessary for correct semantics (these lowerings were
previously applied by `mirgen`).

The procedures in the new `proto_mir` module are largely based on the
procedures from `mirgen`:
* `exprToPMir` is essentially a copy of `mirgen.genx`, but instead of
   MIR being constructed, proto-MIR is constructed
* `isStable` and `isPure` compute the same information as the
  procedures from `mirgen`, just with PMIR as the input
* `classify` is the PMIR's version of `detectKind`