fix compiler crash with macros and enum `$` (#1182)
## Summary

Fix the compiler crashing when returning AST containing a call to the
compiler-generated `$` for an enum type from a macro. The crash only
occurred if the problematic callee is a resolved symbol (not a raw
identifier).

## Details

`sigmatch.initCallCandidate` searches for the module a routine is
owned by, by traversing the owners while skipping generic routines. If
a routine is marked with the `sfFromGeneric` flag, `skipGenericOwner`
unconditionally skips one `owner`. This is because the instantiation of
a generic routine has the generic routine it was instantiated from as
its immediate owner.

The compiler-generated enum-to-string procedure is also marked with
`sfFromGeneric`, but it has the owner of the enum type as the `owner`,
and not a generic routine. If the owner of the enum type is a module
(`skModule`), this meant that `initCallCandidate` skipped over the
`skModule` symbol, eventually reaching the end of the chain and
crashing with an NPE.

Since in the case of `genEnumToStrProc`, the generic procedure
(`EnumToStr` from the `dollars` module) might not be available yet, no
generic symbol can be used as the owner. To fix the problem, the
synthesized procedure is now no longer marked with the `sfFromGeneric`
flag.