sem: `discardable` routines are discarded in `try`
Summary
=======

Routines marked with the `discardable` pragma are now correctly
discarded inside `try`/`except`/`finally` expressions.

Details
=====

Previously the `discardCheck` which handled this did not correctly
traverse `nkTryStmt` when checking for `implicitlyDiscardable`
expressions.

Along with fixing the traversal, such traversals in an error case are no
longer repeated. Instead the traversal is done once, and only once. The
`implicitlyDiscardable` procedure is now simplified to only examine the
input node and no longer attempt traversal, which is now the
responsibility of the caller.

In addition to the fix, a test case has been added to ensure that this
behaviour is preserved, see `trydiscardcheck_try_discardable`.