system: fix `sysFatal` having side effects
Summary
=======

Fix the internal `sysFatal` procedure having side-effects when panics
are enabled (`--panics:on`). This prohibited the usage of procedure
like `rangeCheck` and `raiseInsert` in functions.

This make it possible to bootstrap the compiler with `--panics:on`.

Details
=======

The `sysFatal` procedure immediately terminates the program, so the
access to global state it does perform when panics are enabled
shouldn't be treated as side effects. The problematic side-effects
within `sysFatal` are now casted away.