fix(sem): type-less nodes crashing the compiler (#1186)
## Summary

Fix  `nkEmpty`  and  `nkType`  nodes without a type (which can be
created by a macro) appearing in some type contexts (such as the type of
a conversion) crashing the compiler. A "type expected" error is now
reported.

## Details

`semTypeNode`  assumed that both  `nkEmpty`  and  `nkType`  nodes have a
type assigned already, but this is not true for nodes created by macros.
If the caller of  `semTypeNode`  tried to access the returned type, an
NPE ensued.

It is now verified that both nodes have a type assigned, and if they do
not, a "type expected" user error is reported. An assertion is added to 
`semTypeNode`  to alert about erroneously missing types in the future.

There was one situation where an untyped  `nkEmpty`  node is legal: in
the body slot of a magic type definition such as 
`type set[T]{.magic: "Set".}` .  `typeSectionRightSidePass`  now detects
this case itself and doesn't call  `semTypeNode`  when the body slot
legally contains an  `nkEmpty`  node.