system: fix `sysFatal` having side effects (#1187)
## Summary

Fix the internal `sysFatal` procedure having side-effects when panics
are enabled (`--panics:on`). Usage of procedures like `rangeCheck`
and `raiseInsert` within functions is no longer prevented.

This makes it possible to bootstrap the compiler with `--panics:on`.

## Details

When panics are enabled, the `sysFatal` procedure immediately
terminates the program (no exception is raised), so the access to
global state the implementation does perform shouldn't be treated
as side effects. The problematic side-effects within `sysFatal` are
now cast away.