mir: introduce the proto-MIR (#1184)
## Summary

The proto-MIR is the precursor to the MIR, acting as an intermediate
target for expressions during the AST-to-MIR translation. It is used
for the analysis previously performed on MIR expressions, allowing
`mirgen` to focus more on syntax-related translation.

Semantics don't change, and there there's no deliberate change to how
constructs are translated, although there are some minor differences
(improvements) in the produced MIR.

## Details

### Background

`mirgen` previously performed translation, semantics-related decision
making (e.g., where to inject run-time checks), and attribute
inheritance (e.g., ownership of constructed values) all in one go. For
deciding where and how to inject temporaries, the just-produced MIR was
analyzed (`isPure`, `isStable`, and `detectKind`).

While efficient in terms of memory and time usage, this approach leads
to a fairly complex `mirgen`, with the performed analysis also limited
by the MIR itself.

### The Proto-MIR

The proto-MIR (short: PMIR) is an IR separate from the MIR that only
covers a subset of existing expressions from the AST (calls,
constructions, and complex expressions, are not covered). It represents
expressions in a way such that translating them to the MIR requires as
little conditional logic as possible:
* it includes all lvalue-expression-related run-time checks
* tuple, array, and field access are separated
* `when nimvm` expressions are resolved
* purity of index and dereference operands is already ensured
* some necessary lowering is already applied

#### Data representation

* expressions are represented as *chains* of *items* (not trees of
  nodes)
* the sub-expressions forming an expression (only used for index
  operands at the moment) are part of a single `seq`
* each item stores the originating-from AST node (which still carries
  some information required by the later translation), the type, plus
  item-specific data
* all fields of `ref` type are marked as `.cursor` in order to get
  around the unnecessary ref operations (the objects are guaranteed to
  outlive the PMIR items)

#### Proto-MIR implementation

* translating expression AST to a proto-MIR expression is done via
  `exprToPmir`
* the `exprToMir` implementation is more or less a copy of
  `mirgen.genx`, but with PMIR items being emitted rather than MIR
  nodes
* the `isPure`, `isStable`, and `detectKind` (renamed to `classify`)
  analysis procedures from `mirgen` are moved to the `proto_mir` module
  and adjusted to work with the PMIR
* in contrast to the previous `isPure` implementation, the new one is
  exhaustive and also treats 'let' globals and various rvalue-yielding
  operations as pure

### Integration into `mirgen`

* whenever an expression AST is to be translated, it is first
  translated to PMIR
* the `wantX` routines are called to ensure that the expressions meet
  the contextual requirements (stable, pure, etc.)
* for expressions, `mirgen` only implements the PMIR-to-MIR step (with
  `genx`)

### Minor differences in MIR output

When field checks are disabled, a nested variant access expression such
as `a.f` is now translated to:
```
discard a.case1.case2.case3.f
```
instead of:
```
bind _1 = a.case1
bind _2 = _1.case2
discard _2.case3.f
```

This is because the operand of unchecked variant access is now never
(explicitly) captured, whereas `mirgen.genVariantAccess` previously
always captured the operand for the sake of a simpler implementation.