semexprs: better bracket rewrite reversion
`asBracketExpr` now also detects `[](x, y)` as something to try as a
built-in subscript operation if `x` is a bound generic symbol or symbol
choice.

Not going the `semSubscript` route was previously not a problem, as the
symbol or symbol choice had a type and thus `semOperand` (which reports
an error for un-instantiated generic symbols) wasn't called again, and
so `semOverloadedCall` successfully picked the `mArrGet` overload
(which then call `semSubscript`).

Since the type no longer has a say on whether an operand is analyzed or
not, `asBracketExpr` *must* work properly.