don't emit header includes for unused imported globals (#1191)
## Summary

Imported globals are now only included in the build if they're used
anywhere in alive code, fixing C headers being unnecessarily pulled in
for them.

Closes https://github.com/nim-works/nimskull/issues/1190.
Closes https://github.com/nim-works/nimskull/issues/1125.

## Details

An imported 'let' or 'var' was so far added to the (logical) struct of
the module the global is part of, but this is wrong! The definition of
an imported global doesn't imply the definition of a location owned by
the NimSkull program, rather it makes an external name available with
the given name and type.

**Changes:**
- don't add globals to the module struct if they're marked with
  `sfImportc`
- introduce `symbolToPmir` for querying the PMIR/MIR kind of a symbol
  without adding it to the MIR environment (`nameNode` does this)
- change `mirgen.genLocDef` to only use `nameNode` (which registers the
  symbol with the environment) if the statement constitutes an
  assignment
- add a specification test for the new behaviour

Imported globals are now only registered with the MIR environment (and
subsequently have their definition/include emitted) when they're part
of the alive program.