update the move analyzer
A different strategy is used: each `mnkSink` for a value with custom
copy/sink/destruction behaviour is analyzed for whether it is the last
read of the provide value. If it is, the corresponding data-flow
instruction is changed to an `opConsume` instruction and the sink is
remembered. If it is not, nothing is done.

The difference to the previous one is that:
* the analysis centers around the node kind instead of the DFG
  instruction (this ignoring `mnkConsume` arguments, for which the
  analysis is unnecessary)
* the data-flow graph is modified, making passing the set of sinks-
  turned-into-moves to all further analysis obsolete (less parameters
  and conditionals)

Finally, `computeOwnership` can ignore lvalues with derefs and literals
as the root: the `Sink` modifier is never used for them.