cgen: fix undefined behaviour with object conversion
Summary
=======

Fix up-conversions of `nil` `ref object` or `ptr object` values
invoking undefined behaviour at the C level.

Details
=======

Up conversions of pointers or `ref`s were translated into a
dereference + address-of sequence (e.g., `&a->Sup`), which invokes UB
if the pointer is null.

Instead, the pointer is now directly cast to the super type. According
to the C89 standard (6.5.2.1 "Structure and union specifiers"), "A
pointer to a structure object, suitably converted, points to its
initial member", so this cast is legal.