cgen: fix undefined behaviour of `openArray` conversions (#1197)
## Summary

When in their default empty state, both implicit and explicit
conversions
from  `string`  and  `seq`  to  `openArray`  invoked undefined behaviour
at the
C level. This is fixed now.

Fixes https://github.com/nim-works/nimskull/issues/1194.

## Details

Access of the payload pointer, which is null when the sequence is in
its default state, is now guarded by a not-null check. If the payload
pointer is null, the payload pointer is not accessed and null is used
as the data pointer value for the `openArray` tuple.

Benchmarking this change with the compiler itself showed it to have
little to no effect on execution time.