rework the MIR (part 2) (#1195)
## Summary

This introduces the `copy`, `move`, and `sink` *assignment modifiers*,
which make it explicit where and how ownership is transferred with
assignments. An assignment with an lvalue expression as the source
operand and no modifier is a *shallow copy*, replacing the need for
`fastAsgn`. Previously, the exact semantics of an assignment where an
implicit property of every assignment.

Besides slightly simplifying the `injectdestructors` pass and data-flow
graph construction, the semantics of assignments being properly encoded
in the MIR is a pre-requisite for many MIR-based optimization passes.
More generally, it's a significant step towards decoupling value
ownership semantics from lifetime hooks.

Semantics of existing code don't (intentionally) change, this is an
internal-only rework.

## Details

### MIR

* the `mnkCopy`, `mnkMove`, and `mnkSink` nodes are added, representing
  the modifiers. Syntax-wise, they may only appear directly in
  assignment source slots
* the `mnkFastAsgn` node is obsolete and removed
* `copy` and `move` assignments are *final* . That is, a `copy`
  assignment cannot become a `move` assignment, and the inverse is also
  not possible
* only `sink` assignments are non-final. Collapsing them into either a
  `copy` or `move` assignment is the focus of the move analyzer

### AST-to-MIR Translation

`mirgen` performs the initial placement of assignment modifier, using a
proto-MIR-based analysis. The decision-making works as follows:
1. when the destination is owning (i.e., all non-`.cursor` locals/
   globals/fields):
	1. if a non-destructive move is definitely possible, `move` is used
	2. if a move is never possible (e.g., because the source is a
     cursor), `copy` is used
	3. if whether a move is possible depends on a data-flow analysis,
     `sink` is used
	4. if a destructive move is possible, `sink` is used
	5. if the source expression is an rvalue expression returning an
     owning value (e.g., a call, a construction, etc.), and the
     assignment is an initial assignment, an in-place assignment is
     used
	6. for types with custom copy/sink/destroy behaviour, and non-initial
     assignments, the rvalue is first assigned to a temporary and then
     `move`d into the destination
2. when the destination is non-owning, no modifier is used

Except for the rule 1.6, placement of the modifiers is independent on
whether lifetime hooks are involved. 

In addition, all rvalue expressions (except for `cast`) are now treated
as returning *owning* values. What this means, in effect, is that
temporaries now properly use `def` instead of `def_cursor`. As the
moment, whether `def` or `def_cursor` is used for non-lifetime-hook
using types has no practical effect, but the new behaviour is
technically correct, while the previous one was not.

### Move Analyzer

Operation of the move analyzer changes slightly: instead of looking for
all `opConsume` data-flow instructions, it only looks for those
corresponding to `sink` assignments. Since whether a move or copy is
used is only relevant for lifetime-hook-using types, the `isLastRead`
analysis continues to only be performed for assignments of those types.

Compared to before, the DFG is now updated with the move analyzer
results, removing the need to pass them to each of the following
analysis routine.

### Assignment Rewriting

Rewriting of assignment now only looks for assignments with modifiers,
all assignments without a modifier are ignored. `move`s are exclusively
turned into `=sink` calls (never destructive moves), `copy`s into
`=copy`. `sink` is either turned into a copy, move, or destructive
move. Only the implementation is different, effective program behaviour
doesn't change.

### MIR-to-CGIR Translation

* `move` assignments and *shallow copy* assignments are translated to
  `cnkFastAsgn`
* all other assignments translate to `cnkAsgn`
* the small optimization preventing definitions being moved to the
  start of their scope is expanded to also apply to `def`s of
  temporaries without lifetime hooks, now that `def` is used for
  temporaries more often

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>