fix wrong signed-ness for floats in constants
Summary
=======

Fix the sign of `0.0` float values being inverted in implicit or
explicit constant expression. This affect all backends, but it only
happened spuriously.

Details
=======

`datatables` used `exprStructuralEquivalent`, which treats `0.0` as
being equal to `-0.0`, for comparing nodes storing literal values.
Since the hashing procedure for nodes hashed the float values' bit-
representation, the problem only surfaced in case of hash or bucket
collision.

A dedicated comparison procedure that compares the bit-representations
of float values is now used, fixing the issue.