fix(sem): crash with manual `=destroy` hook call
Summary
=======

Fix a bug where manually calling the `=destroy` (or `=trace`) hook for
a type where the hook is compiler-generated, and the call is within a
procedure definition passed as a typed macro/template parameter that is
part of the macro/template result, crashed the compiler.

Details
=======

All compiler-synthesized hook procedures are marked as `sfFromGeneric`.
The `owner` of such procedures is expected to be the generic procedure
the procedure is an instantiation of (`skipGenericOwner` the immediate
owner), but this wasn't the case for the synthesized hooks, where the
owner was set to the owner of the type the hook is synthesized for.

When a compiler-synthesized hook is used in overload resolution, the
search for the owning module in `initCallCandidate` skips over the
`skModule` symbol (via `skipGenericOwner`), thus resulting in an NPE.

To fix the issue, the owner of synthesized hook procedures is now set
to the hook kinds corresponding generic magic procedure from the
`system` module, making the symbols well-formed.