fix(sem): crash with manual `=destroy` hook call (#1205)
## Summary

Fix a bug where manually calling the `=destroy` (or `=trace`) hook for
a type where the hook is compiler-generated crashed the compiler when
the call is within a typed procedure definition that is returned by a
macro or template.

Fixes https://github.com/nim-works/nimskull/issues/1161.

## Details

All compiler-synthesized hook procedures are marked as `sfFromGeneric`.
The `owner` of such procedures is expected to be the generic procedure
the procedure is an instantiation of (`skipGenericOwner` the immediate
owner), but this wasn't the case for the synthesized hooks, where the
owner was set to the owner of the type the hook is synthesized for.

When a compiler-synthesized hook is used in overload resolution, the
search for the owning module in `initCallCandidate` skips over the
`skModule` symbol (via `skipGenericOwner`), thus resulting in an NPE.

To fix the issue, the owner of synthesized hook procedures is now set
to the hook kind's corresponding generic magic procedure from the
`system` module, making the produced symbols well-formed.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>