mirexec: fix a bug with `traverse`
The `exit` flag was erroneously set to true when the main path was
aborted on a data-flow instruction that's the last instruction in the
subgraph.

Previously, this bug had little to no impact, only affecting the
`needsReset` optimization (unnecessary but otherwise harmless
`wasMoved` calls were injected). For the `injectResultInit` analysis,
the bug led to `result` assignments coming as the last statement within
a body not disabling the default-initialization.

A regression test directly for the `traverse` iterator is added.