mirpasses: handle result initialization with a MIR pass
Instead of using a CGIR-based analysis for figuring out whether the
result variable needs to be default-initialized when using the C
backend, a MIR pass is now used.

Using a MIR pass means that the more powerful MIR data-flow analysis
can be leveraged, allowing for omitting initialization of the result
variable in more cases.

The `cgen`-driven intialization and associated analysis is removed.