cgen: better default initialization for `result` (#1209)
## Summary

Use a MIR pass for handling the default-initialization of `result`
variables for the C backend. This allows using the MIR's more precise
data-flow analysis, allowing for broader omission of `result` variable
initialization.

## Details

* introduce the `injectResultInit` MIR pass
  * it's only used with the C backend for now
  * if a data-flow analysis deems it required, a `result := default()`
    assignment is placed at the start of the body
* remove the `CgNode`-based `result`-assignment analysis from `cgen`

### Data-flow Analysis Bug

A small bug with the `mirexec.traversal` (forward traversal) routine
was discovered and fixed: the `exit` flag was erroneously set to true
when the main path was aborted on a data-flow instruction that's the
last instruction in the subgraph. The `exit` flag must only be true
when the end of the subgraph is reached, so this is wrong.

A regression test for the `traverse` routine is added.

This bug only affected the `injectdestructors.needsReset` optimization
(unnecessary but otherwise harmless `wasMoved` calls were injected
because of it) and `injectResultInit` (unnecessary initialization was
emitted).