fix(mirgen): respect `--panics:on` for overflow checks (#1213)
## Summary

Fix overflow checks for binary integer arithmetic being treated as
potentially raising exceptions when panics are enabled (`--panics:on`).
This didn't affect correctness, but some optimizations were inhibited.

## Details

Use `buildDefectMagicCall` for emitting the binary arithmetic operation
in `mirgen`. If panics are enabled, this makes sure a `mnkCall` is
emitted rather than a `mnkCheckedCall`. 

Checked calls (i.e., calls that can raise an exception) introduce
unstructured control-flow, which, depending on where the call is
located, can result in the compiler having to place additional
`wasMoved` and/or destructor calls.

An `--expandArc`-based test for making sure that all run-time checks
are correctly treated as not raising an exception is added.