cgirgen: produce new-style CGIR
`cgirgen` now emits goto-style CGIR adhering to the structure laid out
in the MIR documentation.

A minor difference compared to the MIR specification is that `cgirgen`
doesn't add potential follow-up targets to `cnkContinueStmt`s. It would
only introduce unnecessary complexity at the moment, the code
generators don't need this information.

In addition, a simple form of eliminating unreachable code is added.
This allows for making sure that `cgen` properly handles the case where
a `finally` is the final jump target of a jump.