jsgen: restore reporting of unhandled exceptions
Instead of keeping track of the enclosing `try` blocks at run-time, for
the purpose of knowing when to report an unhandled exception, module-
level code is now wrapped in a `try`/`catch` that reports escaping
exceptions. This is both simpler to implement *and* has less run-time
overhead.

The `excHandler` global is removed.