jsgen: ensure correct try/finally/catch control-flow
It's possible for the JavaScript `try`s produced for some MIR/CGIR
to not exhibit the required behaviour, as, for example, finally clauses
are entered when they shouldn't be (can happen when a goto jumps over
a finally section).

To prevent this, intermixed with code generation, all jump actions are
analysed for whether the path matches that taken at run-time. If it
wouldn't `catch` and `finally` clauses are flagged as having to be
guarded by boolean flag, which is then toggled prior to a problematic
jump.

Checked calls aren't handled yet.