jsgen: fix yet another enable-flag bug
Consider:
```js
while (true) {
  L1: {
    try {
      if (...) {
        var Enabled1_ = true;
        Enabled1_ = false;
        break L1;
      } else {
        break L1;
      }
    } finally {
      if (Enabled1_ === undefined || Enabled_1) {
        ...
      }
    }
  }
}
```
This code won't behave as expected, as the second break doesn't execute
the finally. The flag needs to be set back to true on exit of the
section.

The naming change from earlier could be reverted again, but using
distinct names makes the code a bit clearer, so it's kept.