jsgen: fix missing current exception restoration
A small oversight. Since the current exception is restored when
*landing* (instead of when *leaving*), it's possible that an exception
handler within a procedure is exited through raising another exception,
but without there being an intercepting finally that restores the
exception. Therefore, all exception control-flow leaving procedures
needs to be treated as requiring a restore on landing.

It's not possible to add a test for this behaviour, since it's at odds
with the behaviour of the C and VM backend.