mirgen: implement "never raises" handling
If a procedure is flagged with `sfNeverRaises`, the body is wrapped in
a catch-all exception handler, where the new `nimUnhandledException`
runtime procedure is invoked from the handler. If the handler is not
actually used, `cgirgen`'s unreachable code elimination automatically
eliminates it again.

Implementing this in `mirgen` ensures consistent behaviour across all
backends.