std/dynlib: fix `loadLibPatterns` having unknown effects (#1237)
## Summary

The `loadLibPatterns` procedure was erroneously inferred to have
unknown effects (`RootEffect`), making it unusable in `.tags: []`
procedures. This is now fixed.

## Details

`loadLibPatterns` calls the at-that-point forwarded procedure
`loadLib`, so effect inference falls back to assume that
`loadLibPatterns` can have effects.

Since `loadLib`, as well as the other forward-declared procedures,
don't have effects in practice, this is now explicitly specified
on the forward declarations.