std/macros: const for nodes requring special init
Summary
========

Introduce a constant `nnkRequireInitKinds` for all nodes that require
special initialization, and therefore cannot be constructed via
`newNimNode`. This allows imports of `std/macros` that construct
arbitrary nodes to create a filter.

Details
======

Added the constant to `std/macros`, with export, as well as a test to
ensure we can detect any node kinds not in the list in `tmacros`.