internal: fix an incorrect assertion
Summary
=======

An assertion in the `injectdestructors` module was incorrect, which led
to the compiler aborting for `x.a = f()` assignments where `a` is a
discriminator field.

Details
=======

Only raw lvalue expressions must no appear as the source operand of a
branch-switch assignment, rvalue expression (which aren't included in
`ModifierNodes`) are okay.