implement `nimUnhandledException` for the VM
The procedure is implemented as a syscall (`vmops`), which raises the
existing `vmEvtUnhandledException` VM event, thus terminating the VM
right away.

Since syscall implementations are currently unable to create a stack-
trace, no stack-trace is passed along, which
`compilerbridge.buildError` now accounts for (by creating the stack-
trace itself). This is somewhat hack-y.